adding personal group expenses and budget feature

+++ src/main/java/com/backend/senior_backend/dto/budgetAndExpensesDTO.java
@@ -8,6 +8,6 @@ import lombok.Data;
 public class budgetAndExpensesDTO {
 
     final int budget;
-    final int expenses;
+    final Double expenses;
 
 }
